no major changes

+++ app/src/main/java/sketch/components/Line.java
@@ -12,14 +12,14 @@ public class Line extends Drawable {
         this.end = end;
     }
 
-    public Line(float startX, float startY, float endX, float endY) {
+    public Line(double startX, double startY, double endX, double endY) {
         this.start = new Coordinate(startX, startY);
         this.end = new Coordinate(endX, endY);
     }
     
     @Override
     public void draw(Graphics2D g2d) {
-        g2d.draw(new Line2D.Float(this.start.x, this.start.y, this.end.x, this.end.y));
+        g2d.draw(new Line2D.Double(this.start.x, this.start.y, this.end.x, this.end.y));
     }
 }
